refactor(gmail): reorganize Gmail provider implementation into separate module
- Moved the Gmail provider implementation from `mod.rs` to a new `provider.rs` file, enhancing code organization and maintainability.
- Retained the incremental sync logic and user profile fetching functionality, ensuring no change in behavior.
- Updated module imports accordingly to reflect the new structure.
